refactor: move assign state logic to a hook, to be reusable
